==> 3_SP_results.ipynb <==
import pickle
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns

def create_results_df():
    # Load pickled scores for each model
    results_dict = pickle.load(open("model_scores.p", "rb"))

    # Create pandas df and save as csv
    results_df = pd.DataFrame.from_dict(results_dict, orient='index',columns=['RMSE', 'MAE', 'R2'])
    results_df = results_df.sort_values(by='RMSE',ascending=False).reset_index()

    results_df.to_csv('results.csv')

    return results_df

def plot_results(results_df):
    fig, ax = plt.subplots(figsize=(12, 5))
    sns.lineplot(x=np.arange(len(results_df)), y='RMSE', data=results_df, ax=ax,label='RMSE', color='mediumblue')
    sns.lineplot(x=np.arange(len(results_df)), y='MAE', data=results_df, ax=ax,label='MAE', color='Cyan')

    plt.xticks(np.arange(len(results_df)), rotation=45)
    ax.set_xticklabels(results_df['index'])
    ax.set(xlabel="Model",ylabel="Scores",title="Model Error Comparison")
    sns.despine()

    plt.savefig(f'compare_models.png')

results = create_results_df()
plot_results(results)
