==> 4_SP_newdataupdate.ipynb <==
import pandas as pd

o_df=pd.read_csv('sales.csv')
n_df=pd.read_csv('data_2018.csv') #enter new dataset name everytime while uploading new data

o_df=o_df[['date', 'sales']]
o_df.head()

o_df.tail()

n_df=n_df[['date', 'sales']]
n_df.head()

n_df.tail()

merge_df = pd.concat([o_df, n_df])
merge_df.head()

merge_df.tail()

merge_df.to_csv('sales.csv')
